+++ homework-3/src/main/resources/liquibase/changelogs/homework-2/sql/05.create-customers-accounts-table.sql
@@ -1,8 +1,8 @@
 CREATE TABLE banking.customers_accounts
 (
-    id          SERIAL PRIMARY KEY,
-    customer_id INT NOT NULL,
-    account_id  INT NOT NULL,
+    id          BIGSERIAL PRIMARY KEY,
+    customer_id BIGINT NOT NULL,
+    account_id  BIGINT NOT NULL,
     FOREIGN KEY (customer_id) REFERENCES banking.customers (id),
     FOREIGN KEY (account_id) REFERENCES banking.accounts (id)
 );
